+++ src/components/s3/bucketList/dialogs/index.ts
@@ -1,0 +1,1 @@
+export { default as CreateBucketDialog } from "./CreateBucketDialog";
